Fixed calculation of true anomaly from position vectors which was incorrectly using ship orbital velocity.
I also identified an issue using the dot product to determine when to negate true since the dot product for position & velocity are perpendicular in some cases (especially for circular orbits).  I solved this by comparing the cross product of the position & eccentricity vector with the cross product of the orbits momentum instead.